% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_age.R
\name{give_age}
\alias{give_age}
\title{Give an age}
\usage{
give_age(unit = "years")
}
\arguments{
\item{unit}{string, either "years" or "months"}
}
\value{
An integer
}
\description{
Give an age
}
\examples{
give_age()
give_age(unit = "months")
}
