% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_name.R
\name{give_name}
\alias{give_name}
\title{Give a name}
\usage{
give_name(sex = "female")
}
\arguments{
\item{sex}{string, either male or female}
}
\value{
string
}
\description{
Give a name
}
\examples{
give_name()
give_name(sex = "male")
}
