% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_names.R
\docType{data}
\name{sa_names}
\alias{sa_names}
\title{Names given to babies born i 2016 in South Africa}
\format{a data.frame with columns first_name number, sex.}
\source{
Stats SA
}
\description{
data about live births collected by Stats SA, see \url{http://www.statssa.gov.za/?page_id=1854&PPN=P0305&SCH=7092}.
The data is distributed with a CC-BY SA license.
}
