% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someone-package.R
\docType{package}
\name{someone-package}
\alias{someone}
\alias{someone-package}
\title{someone: Creates fake people}
\description{
Creates fake people data such as birth year and name.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maelle/someone}
  \item Report bugs at \url{https://github.com/maelle/someone/issues}
}

}
\author{
\strong{Maintainer}: Maëlle Salmon \email{maelle.salmon@yahoo.se}

}
\keyword{internal}
